voice-ingestion/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.0
python-dotenv==1.0.1
openai==1.54.0
httpx==0.27.0
sqlalchemy==2.0.44
